requirements.txt
allure-pytest==2.13.3
allure-python-commons==2.13.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
execnet==2.0.2
Faker==24.3.0
idna==3.6
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
